+++ library/src/test/java/com/anthonycr/bonsai/SchedulersUnitTest.java
@@ -20,9 +20,8 @@
  */
 package com.anthonycr.bonsai;
 
+import android.os.Handler;
 import android.os.Looper;
-
-import junit.framework.Assert;
 
 import org.junit.Test;
 
@@ -31,7 +30,9 @@ import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicReference;
 
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 public class SchedulersUnitTest extends BaseUnitTest {
@@ -57,48 +58,28 @@ public class SchedulersUnitTest extends BaseUnitTest {
     }
 
     @Test
-    public void testCurrentScheduler_initializesLooper() throws Exception {
-        final AtomicReference<Boolean> currentScheduler = new AtomicReference<>();
-        final AtomicReference<Boolean> nullScheduler = new AtomicReference<>();
+    public void testImmediateScheduler_isSynchronous() throws Exception {
+        final AtomicReference<Boolean> executedBoolean = new AtomicReference<>(false);
+        final AtomicReference<Looper> currentScheduler = new AtomicReference<>();
         final CountDownLatch latch = new CountDownLatch(1);
-        Schedulers.worker().execute(new Runnable() {
-            @Override
-            public void run() {
-                // should be null
-                Looper currentLooper = Looper.myLooper();
-                nullScheduler.set(currentLooper == null);
-                Schedulers.current().execute(new Runnable() {
-                    @Override
-                    public void run() {
-                        // now it shouldn't be null
-                        currentScheduler.set(Looper.myLooper() != null);
-                    }
-                });
-                latch.countDown();
-            }
-        });
-        latch.await();
-        Assert.assertTrue(nullScheduler.get());
-        Assert.assertTrue(currentScheduler.get());
-    }
 
-    @Test
-    public void testCurrentScheduler_isCorrect() throws Exception {
-        final AtomicReference<Boolean> currentScheduler = new AtomicReference<>();
-        final CountDownLatch latch = new CountDownLatch(1);
         Utils.prepareLooper();
         final Looper currentLooper = Looper.myLooper();
-        Schedulers.current().execute(new Runnable() {
+
+        Schedulers.immediate().execute(new Runnable() {
             @Override
             public void run() {
                 Utils.prepareLooper();
-                currentScheduler.set(Looper.myLooper() == currentLooper);
+                currentScheduler.set(Looper.myLooper());
+                executedBoolean.set(true);
                 latch.countDown();
             }
         });
 
+        assertTrue(executedBoolean.get());
+
         latch.await();
-        assertTrue(currentScheduler.get());
+        assertEquals(currentScheduler.get(), currentLooper);
     }
 
     @Test
@@ -156,6 +137,36 @@ public class SchedulersUnitTest extends BaseUnitTest {
 
         latch.await();
         assertFalse(currentScheduler.get());
+    }
+
+    @Test
+    public void testFromHandler_isOnRightThread() throws Exception {
+        final AtomicReference<String> executorThread = new AtomicReference<>(null);
+        final AtomicReference<String> schedulerThread = new AtomicReference<>(null);
+        final CountDownLatch latch = new CountDownLatch(1);
+        Executors.newSingleThreadExecutor()
+            .execute(new Runnable() {
+                @Override
+                public void run() {
+                    Looper.prepare();
+                    Handler handler = new Handler(Looper.myLooper());
+                    executorThread.set(Thread.currentThread().toString());
+                    Schedulers.from(handler)
+                        .execute(new Runnable() {
+                            @Override
+                            public void run() {
+                                schedulerThread.set(Thread.currentThread().toString());
+                                latch.countDown();
+                            }
+                        });
+                }
+            });
+
+        latch.await();
+
+        assertNotNull(schedulerThread.get());
+        assertNotNull(executorThread.get());
+        assertEquals(schedulerThread.get(), executorThread.get());
     }
 
     @Test